@@ -5,27 +5,35 @@
   "description": "Drd+ tables and their units conversion",
   "autoload": {
     "psr-4": {
-      "DrdPlus\\Tables\\": "Tables"
+      "DrdPlus\\Tables\\": "src"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "DrdPlus\\Tests\\": "tests"
+      "DrdPlus\\Tests\\": "tests",
+      "Granam\\Tests\\ExceptionsHierarchy\\": "vendor/granam/exceptions-hierarchy/tests/ExceptionsHierarchy"
     }
   },
   "require": {
-    "php": ">=7.3",
-    "granam/float": "^5.0",
-    "granam/boolean": "^3.2",
-    "drdplus/base-properties": "^2.1",
+    "php": ">=7.4",
+    "drdplus/base-properties": "^3.0",
     "drdplus/calculations": "^1.0",
-    "granam/dice-rolls": "^6.0"
+    "drdplus/codes": "^2.3",
+    "granam/boolean": "^3.2",
+    "granam/dice-rolls": "^6.0",
+    "granam/float": "^5.0",
+    "granam/integer": "^7.1",
+    "granam/number": "^5.1",
+    "granam/strict-object": "^3.1",
+    "granam/string": "^4.1",
+    "granam/tools": "^6.1"
   },
   "require-dev": {
     "granam/exceptions-hierarchy": "~5.0",
     "granam/test-with-mockery": "^2.0",
     "mockery/mockery": "~1.2",
     "phpunit/phpunit": "~9.3",
+    "rector/rector": "^0.11.53",
     "roave/security-advisories": "dev-master"
   },
   "config": {
