@@ -94,7 +94,7 @@ class SkillsByBackgroundPointsTableTest extends TableTest
         return $combinations;
     }
 
-    private static $expectedSkillPoints = [
+    private static array $expectedSkillPoints = [
         [1, 1, 1, 2, 0, 1, 1, 1, 1, 2, 0, 1, 0, 3, 0, 0, 2, 1, 0, 1, 2,],
         [1, 1, 1, 3, 0, 1, 2, 1, 1, 2, 0, 2, 1, 3, 0, 0, 3, 1, 0, 2, 2,],
         [1, 1, 1, 4, 0, 1, 2, 1, 2, 3, 0, 2, 1, 4, 0, 0, 4, 1, 1, 2, 2,],
