@@ -1,4 +1,4 @@
-<?php declare(strict_types = 1);
+<?php
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
